+++ .oh/cli/legacy/bin/oh.js
@@ -1,0 +1,2 @@
+#!/usr/bin/env node
+import "@mifune/agro/dist/agro.js";